Pratham_ArgyleEnigma_Assignment/src/backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
openai
pytest
python-dotenv
pytesseract
python-dotenv
google-generativeai
